refactor: reduce duplication, separate utils
